requirements.txt
altair==5.5.0
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.0
attrs==25.3.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.4.0
cachetools==5.5.2
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dotenv==0.9.9
et_xmlfile==2.0.0
Flask==3.1.0
flatbuffers==25.9.23
fonttools==4.57.0
fsspec==2025.3.2
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.43.0
google-genai==1.53.0
googlemaps==4.10.0
greenlet==3.3.0
h11==0.16.0
h3==4.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.36.0
numpy==2.4.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
pillow==12.1.0
playwright==1.57.0
plotly==6.0.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
pyodbc==5.2.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
PySocks==1.7.1
pyswisseph==2.10.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.2
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.34.0
setuptools==80.10.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
streamlit==1.44.1
tenacity==9.1.2
tennis-betting @ file:///C:/Users/pncdp/OneDrive/VFB%20Group/Gambling%20Fortuna/Betting%20Tennis/dist/tennis_betting-0.1.0-py3-none-any.whl#sha256=5d49545c89b83c3095e9d30c741204e983991bc13c5906dd0fa3ed6a02c46212
threadpoolctl==3.6.0
tiktoken==0.12.0
timezonefinder==8.1.0
toml==0.10.2
tornado==6.4.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.46.3
wsproto==1.2.0
